create methods cleaned up
